Uso función de redondeo en vez de pasar a single

+++ t_estabilidad.m
@@ -1,6 +1,6 @@
 function Te = t_estabilidad (P1, P2, alpha, beta, k, j)
-  yd = serie(P1, P2, alpha, beta, k, j, "double")
-  ys = serie(P1, P2, alpha, beta, k, j, "single")
-  mu_s = unidadmaquina(2, "single")
+  yd = double(serie(P1, P2, alpha, beta, k, j, "double"))
+  ys = double(serie(P1, P2, alpha, beta, k, j, "single"))
+  mu_s = double(unidadmaquina(2, "single"))
   Te = abs(yd-ys)/(abs(yd)*mu_s);
 end
